Move existing tests to correct namespace
